feat: Added some parameters to to_sql methods
Added some parameters to pandas.DataFrame.to_sql methods to store data
into SQLite.

- if_exists: How to behave if the table already exists.
- chunksize: Rows will be written in batches of this size at a time.

When it is not provided a chunksize, the full result of the query is
put in a dataframe at once. Sometimes the SQLite driver or the system
might have a limit to this, and if the size of the data is more than
the limit, you will get an error.